move the map into openapi_path_mappings.go
